correction du contage des variables ajout d'un devin ::p

+++ PM.h
@@ -1,7 +1,7 @@
 #ifndef _PM_H_
 #define _PM_H_
 
-double factorielle(int n);
+long double factorielle(int n);
 double bazinga();
 double distance (double x1, double y1, double x2, double y2);
 void intro ();
